Move indexing-related stuff to Indexer
from MetalsLanguageServer
