+++ metals/src/main/scala/scala/meta/internal/metals/MetalsLanguageServer.scala
@@ -7,18 +7,12 @@ import java.nio.file._
 import java.util
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.Executors
-import java.util.concurrent.ForkJoinPool
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.atomic.AtomicReference
-import java.{util => ju}
 
 import scala.collection.immutable.Nil
-import scala.collection.mutable
-import scala.collection.mutable.ArrayBuffer
-import scala.collection.parallel.CollectionConverters._
-import scala.collection.parallel.ForkJoinTaskSupport
 import scala.concurrent.Await
 import scala.concurrent.ExecutionContextExecutorService
 import scala.concurrent.Future
@@ -30,7 +24,6 @@ import scala.util.Failure
 import scala.util.Success
 import scala.util.control.NonFatal
 
-import scala.meta.dialects._
 import scala.meta.internal.bsp.BspConfigGenerationStatus._
 import scala.meta.internal.bsp.BspConfigGenerator
 import scala.meta.internal.bsp.BspConnector
@@ -42,7 +35,6 @@ import scala.meta.internal.builds.BuildServerProvider
 import scala.meta.internal.builds.BuildTool
 import scala.meta.internal.builds.BuildToolSelector
 import scala.meta.internal.builds.BuildTools
-import scala.meta.internal.builds.Digest.Status
 import scala.meta.internal.builds.NewProjectProvider
 import scala.meta.internal.builds.ShellRunner
 import scala.meta.internal.builds.WorkspaceReload
@@ -82,7 +74,6 @@ import scala.meta.internal.parsing.TokenEditDistance
 import scala.meta.internal.parsing.Trees
 import scala.meta.internal.remotels.RemoteLanguageServer
 import scala.meta.internal.rename.RenameProvider
-import scala.meta.internal.semanticdb.Scala._
 import scala.meta.internal.semver.SemVer
 import scala.meta.internal.tvp._
 import scala.meta.internal.worksheets.DecorationWorksheetPublisher
@@ -780,7 +771,7 @@ class MetalsLanguageServer(
             languageClient,
             buildClient,
             () => userConfig,
-            () => profiledIndexWorkspace(() => ()),
+            () => indexer.profiledIndexWorkspace(() => ()),
             () => workspace,
             () => focusedDocument,
             buildTargets,
@@ -1395,7 +1386,7 @@ class MetalsLanguageServer(
     Future
       .sequence(
         List(
-          Future(reindexWorkspaceSources(paths)),
+          Future(indexer.reindexWorkspaceSources(paths)),
           compilations.compileFiles(paths),
           onBuildChanged(paths).ignoreValue
         ) ++ paths.map(f => Future(interactiveSemanticdbs.textDocument(f)))
@@ -1687,7 +1678,7 @@ class MetalsLanguageServer(
     params match {
       case ServerCommands.ScanWorkspaceSources() =>
         Future {
-          indexWorkspaceSources()
+          indexer.indexWorkspaceSources()
         }.asJavaObject
       case ServerCommands.RestartBuildServer() =>
         bspSession.foreach { session =>
@@ -2135,7 +2126,7 @@ class MetalsLanguageServer(
             case Some(digest) if isBloopOrEmpty =>
               slowConnectToBloopServer(forceImport, buildTool, digest)
             case Some(digest) =>
-              reloadWorkspaceAndIndex(forceImport, buildTool, digest)
+              indexer.reloadWorkspaceAndIndex(forceImport, buildTool, digest)
           }
         case None =>
           Future.successful(BuildChange.None)
@@ -2279,7 +2270,7 @@ class MetalsLanguageServer(
         buildTargets.resetConnections(idToConnection)
         lastImportedBuilds = bspBuilds.map(_.build)
       }
-      _ <- profiledIndexWorkspace(() => doctor.check())
+      _ <- indexer.profiledIndexWorkspace(() => doctor.check())
       _ = if (session.main.isBloop) checkRunningBloopVersion(session.version)
       _ = diagnostics.reset()
     } yield {
@@ -2287,246 +2278,43 @@ class MetalsLanguageServer(
     }
   }
 
-  private def indexWorkspaceSources(): Unit = {
-    case class SourceToIndex(
-        source: AbsolutePath,
-        sourceItem: AbsolutePath,
-        targets: Iterable[b.BuildTargetIdentifier]
-    )
-    val sourcesToIndex = mutable.ListBuffer.empty[SourceToIndex]
-    for {
-      (sourceItem, targets) <- buildTargets.sourceItemsToBuildTargets
-      source <- sourceItem.listRecursive
-      if source.isScalaOrJava
-    } {
-      targets.asScala.foreach(buildTargets.linkSourceFile(_, source))
-      sourcesToIndex += SourceToIndex(source, sourceItem, targets.asScala)
-    }
-    val threadPool = new ForkJoinPool(
-      Runtime.getRuntime().availableProcessors() match {
-        case 1 => 1
-        case f => f / 2
-      }
-    )
-    try {
-      val parSourcesToIndex = sourcesToIndex.toSeq.par
-      parSourcesToIndex.tasksupport = new ForkJoinTaskSupport(threadPool)
-      parSourcesToIndex.foreach(f =>
-        indexSourceFile(f.source, Some(f.sourceItem), f.targets.headOption)
-      )
-    } finally threadPool.shutdown()
-  }
-
-  private def reindexWorkspaceSources(
-      paths: Seq[AbsolutePath]
-  ): Unit = {
-    for {
-      path <- paths.iterator
-      if path.isScalaOrJava
-    } {
-      indexSourceFile(path, buildTargets.inverseSourceItem(path), None)
-    }
-  }
-
-  private def sourceToIndex(
-      source: AbsolutePath,
-      targetOpt: Option[b.BuildTargetIdentifier]
-  ): AbsolutePath =
-    targetOpt
-      .flatMap(ammonite.generatedScalaPath(_, source))
-      .getOrElse(source)
-
-  private def indexSourceFile(
-      source: AbsolutePath,
-      sourceItem: Option[AbsolutePath],
-      targetOpt: Option[b.BuildTargetIdentifier]
-  ): Unit = {
-
-    try {
-      val sourceToIndex0 = sourceToIndex(source, targetOpt)
-      if (sourceToIndex0.exists) {
-        val dialect = {
-          val scalaVersion =
-            targetOpt
-              .flatMap(buildTargets.scalaTarget)
-              .map(_.scalaVersion)
-              .getOrElse(
-                scalaVersionSelector.fallbackScalaVersion(
-                  source.isAmmoniteScript
-                )
-              )
-          ScalaVersions.dialectForScalaVersion(
-            scalaVersion,
-            includeSource3 = true
-          )
-        }
-        val reluri = source.toIdeallyRelativeURI(sourceItem)
-        val input = sourceToIndex0.toInput
-        val symbols = ArrayBuffer.empty[WorkspaceSymbolInformation]
-        SemanticdbDefinition.foreach(input, dialect) {
-          case SemanticdbDefinition(info, occ, owner) =>
-            if (WorkspaceSymbolProvider.isRelevantKind(info.kind)) {
-              occ.range.foreach { range =>
-                symbols += WorkspaceSymbolInformation(
-                  info.symbol,
-                  info.kind,
-                  range.toLSP
-                )
-              }
-            }
-            if (
-              sourceItem.isDefined &&
-              !info.symbol.isPackage &&
-              (owner.isPackage || source.isAmmoniteScript)
-            ) {
-              definitionIndex.addToplevelSymbol(
-                reluri,
-                source,
-                info.symbol,
-                dialect
-              )
-            }
-        }
-        workspaceSymbols.didChange(source, symbols.toSeq)
-
-        // Since the `symbols` here are toplevel symbols,
-        // we cannot use `symbols` for expiring the cache for all symbols in the source.
-        symbolDocs.expireSymbolDefinition(sourceToIndex0, dialect)
-      }
-    } catch {
-      case NonFatal(e) =>
-        scribe.error(source.toString(), e)
-    }
-  }
-
-  private def profiledIndexWorkspace(check: () => Unit): Future[Unit] = {
-    val tracked = statusBar.trackFuture(
-      s"Indexing",
-      Future {
-        timerProvider.timedThunk("indexed workspace", onlyIf = true) {
-          try indexWorkspace(check)
-          finally {
-            Future(scalafixProvider.load())
-            indexingPromise.trySuccess(())
-          }
-        }
-      }
-    )
-    tracked.foreach { _ =>
-      statusBar.addMessage(
-        s"${clientConfig.icons.rocket}Indexing complete!"
-      )
-      if (clientConfig.initialConfig.statistics.isMemory) {
-        logMemory(
-          "definition index",
-          definitionIndex
-        )
-        logMemory(
-          "references index",
-          referencesProvider.index
-        )
-        logMemory(
-          "workspace symbol index",
-          workspaceSymbols.inWorkspace
-        )
-        logMemory(
-          "classpath symbol index",
-          workspaceSymbols.inDependencies.packages
-        )
-        logMemory(
-          "build targets",
-          buildTargets
-        )
-      }
-    }
-    tracked
-  }
-
-  private def logMemory(name: String, index: Object): Unit = {
-    val footprint = Memory.footprint(name, index)
-    scribe.info(s"memory: $footprint")
-  }
-
   private var lastImportedBuilds = List.empty[ImportedBuild]
 
-  private def indexWorkspace(check: () => Unit): Unit = {
-    val i = (ammonite.lastImportedBuild :: lastImportedBuilds).reduce(_ ++ _)
-    timerProvider.timedThunk(
-      "updated build targets",
-      clientConfig.initialConfig.statistics.isIndex
-    ) {
-      buildTargets.reset()
-      interactiveSemanticdbs.reset()
-      buildClient.reset()
-      semanticDBIndexer.reset()
-      treeView.reset()
-      worksheetProvider.reset()
-      symbolSearch.reset()
-      buildTargets.addWorkspaceBuildTargets(i.workspaceBuildTargets)
-      buildTargets.addScalacOptions(i.scalacOptions)
-      buildTargets.addJavacOptions(i.javacOptions)
-      for {
-        item <- i.sources.getItems.asScala
-        source <- item.getSources.asScala
-      } {
-        buildTargets.addSourceItem(source, item.getTarget)
-      }
-      check()
-      buildTools
-        .loadSupported()
-      formattingProvider.validateWorkspace()
-    }
-    timerProvider.timedThunk(
-      "started file watcher",
-      clientConfig.initialConfig.statistics.isIndex
-    ) {
-      try {
-        fileWatcher.restart()
-      } catch {
-        // note(@tgodzik) This is needed in case of ammonite
-        // where it can rarely deletes directories while we are trying to watch them
-        case NonFatal(e) =>
-          scribe.warn("File watching failed, indexes will not be updated.", e)
-      }
-    }
-    timerProvider.timedThunk(
-      "indexed library classpath",
-      clientConfig.initialConfig.statistics.isIndex
-    ) {
-      workspaceSymbols.indexClasspath()
-    }
-    timerProvider.timedThunk(
-      "indexed workspace SemanticDBs",
-      clientConfig.initialConfig.statistics.isIndex
-    ) {
-      semanticDBIndexer.onTargetRoots()
-    }
-    timerProvider.timedThunk(
-      "indexed workspace sources",
-      clientConfig.initialConfig.statistics.isIndex
-    ) {
-      indexWorkspaceSources()
-    }
-    timerProvider.timedThunk(
-      "indexed library sources",
-      clientConfig.initialConfig.statistics.isIndex
-    ) {
-      indexDependencySources(i.dependencySources)
-    }
-
-    focusedDocument.foreach { doc =>
-      buildTargets
-        .inverseSources(doc)
-        .foreach(focusedDocumentBuildTarget.set)
-    }
-
-    val targets = buildTargets.allBuildTargetIds
-    buildTargetClasses
-      .rebuildIndex(targets)
-      .foreach { _ =>
-        languageClient.refreshModel()
-      }
-  }
+  private val indexer = Indexer(
+    () => workspaceReload,
+    () => doctor,
+    languageClient,
+    () => bspSession,
+    executionContext,
+    () => tables,
+    () => statusBar,
+    timerProvider,
+    () => scalafixProvider,
+    indexingPromise,
+    () => ammonite,
+    () => lastImportedBuilds,
+    clientConfig,
+    definitionIndex,
+    () => referencesProvider,
+    () => workspaceSymbols,
+    buildTargets,
+    () => interactiveSemanticdbs,
+    () => buildClient,
+    () => semanticDBIndexer,
+    () => treeView,
+    () => worksheetProvider,
+    () => symbolSearch,
+    () => buildTools,
+    () => formattingProvider,
+    fileWatcher,
+    () => focusedDocument,
+    focusedDocumentBuildTarget,
+    buildTargetClasses,
+    () => userConfig,
+    sh,
+    symbolDocs,
+    scalaVersionSelector
+  )
 
   private def checkRunningBloopVersion(bspServerVersion: String) = {
     if (doctor.isUnsupportedBloopVersion()) {
@@ -2546,89 +2334,6 @@ class MetalsLanguageServer(
           case _ =>
         }
       }
-    }
-  }
-
-  private def indexDependencySources(
-      dependencySources: b.DependencySourcesResult
-  ): Unit = {
-    // Track used Jars so that we can
-    // remove cached symbols from Jars
-    // that are not used
-    val usedJars = mutable.HashSet.empty[AbsolutePath]
-    val jdkSources = JdkSources(userConfig.javaHome)
-    jdkSources match {
-      case Right(zip) =>
-        usedJars += zip
-        addSourceJarSymbols(zip)
-      case Left(notFound) =>
-        val candidates = notFound.candidates.mkString(", ")
-        scribe.warn(
-          s"Could not find java sources in $candidates. Java symbols will not be available."
-        )
-    }
-    val isVisited = new ju.HashSet[String]()
-    for {
-      item <- dependencySources.getItems.asScala
-      _ = jdkSources.foreach(source =>
-        buildTargets.addDependencySource(source, item.getTarget)
-      )
-      sourceUri <- Option(item.getSources).toList.flatMap(_.asScala)
-      path = sourceUri.toAbsolutePath
-      _ = buildTargets.addDependencySource(path, item.getTarget)
-      if !isVisited.contains(sourceUri)
-    } {
-      isVisited.add(sourceUri)
-      try {
-        if (path.isJar) {
-          usedJars += path
-          addSourceJarSymbols(path)
-        } else if (path.isDirectory) {
-          val dialect = buildTargets
-            .scalaTarget(item.getTarget)
-            .map(scalaTarget =>
-              ScalaVersions.dialectForScalaVersion(
-                scalaTarget.scalaVersion,
-                includeSource3 = true
-              )
-            )
-            .getOrElse(Scala213)
-          definitionIndex.addSourceDirectory(path, dialect)
-        } else {
-          scribe.warn(s"unexpected dependency: $path")
-        }
-      } catch {
-        case NonFatal(e) =>
-          scribe.error(s"error processing $sourceUri", e)
-      }
-    }
-    // Schedule removal of unused toplevel symbols from cache
-    sh.schedule(
-      new Runnable {
-        override def run(): Unit = {
-          tables.jarSymbols.deleteNotUsedTopLevels(usedJars.toArray)
-        }
-      },
-      2,
-      TimeUnit.SECONDS
-    )
-  }
-
-  /**
-   * Add top level Scala symbols from source JAR into symbol index
-   * Uses H2 cache for symbols
-   *
-   * @param path JAR path
-   */
-  private def addSourceJarSymbols(path: AbsolutePath): Unit = {
-    tables.jarSymbols.getTopLevels(path) match {
-      case Some(toplevels) =>
-        val dialect = ScalaVersions.dialectForDependencyJar(path.filename)
-        definitionIndex.addIndexedSourceJar(path, toplevels, dialect)
-      case None =>
-        val dialect = ScalaVersions.dialectForDependencyJar(path.filename)
-        val toplevels = definitionIndex.addSourceJar(path, dialect)
-        tables.jarSymbols.putTopLevels(path, toplevels)
     }
   }
 
@@ -2657,60 +2362,6 @@ class MetalsLanguageServer(
       slowConnectToBuildServer(forceImport = false)
     } else {
       Future.successful(BuildChange.None)
-    }
-  }
-
-  private def reloadWorkspaceAndIndex(
-      forceRefresh: Boolean,
-      buildTool: BuildTool,
-      checksum: String
-  ): Future[BuildChange] = {
-    def reloadAndIndex(session: BspSession): Future[BuildChange] = {
-      workspaceReload.persistChecksumStatus(Status.Started, buildTool)
-
-      session
-        .workspaceReload()
-        .map { _ =>
-          scribe.info("Correctly reloaded workspace")
-          profiledIndexWorkspace(() => doctor.check())
-          workspaceReload.persistChecksumStatus(Status.Installed, buildTool)
-          BuildChange.Reloaded
-        }
-        .recoverWith { case NonFatal(e) =>
-          scribe.error(s"Unable to reload workspace: ${e.getMessage()}")
-          workspaceReload.persistChecksumStatus(Status.Failed, buildTool)
-          languageClient.showMessage(Messages.ReloadProjectFailed)
-          Future.successful(BuildChange.Failed)
-        }
-    }
-
-    bspSession match {
-      case None =>
-        scribe.warn("No build session currently active to reload.")
-        Future.successful(BuildChange.None)
-      case Some(session) if forceRefresh => reloadAndIndex(session)
-      case Some(session) =>
-        workspaceReload.oldReloadResult(checksum) match {
-          case Some(status) =>
-            scribe.info(s"Skipping reload with status '${status.name}'")
-            Future.successful(BuildChange.None)
-          case None =>
-            for {
-              userResponse <- workspaceReload.requestReload(
-                buildTool,
-                checksum
-              )
-              installResult <- {
-                if (userResponse.isYes) {
-                  reloadAndIndex(session)
-                } else {
-                  tables.dismissedNotifications.ImportChanges
-                    .dismiss(2, TimeUnit.MINUTES)
-                  Future.successful(BuildChange.None)
-                }
-              }
-            } yield installResult
-        }
     }
   }
